change inixTimestamp to long

+++ disability-map/Dtos/ReservationDto/GetReservationByPlace.cs
@@ -5,7 +5,7 @@ namespace disability_map.Dtos
     public class GetReservationByPlace
     {
         public UserContactDto User{ get; set; }
-        public int UnixTimestamp { get; set; }
+        public long UnixTimestamp { get; set; }
 
         public long seq { get; set; }
     }
