+++ disability-map/Dtos/ReservationDto/GetReservationByUser.cs
@@ -5,7 +5,7 @@ namespace disability_map.Dtos
     public class GetReservationByUser
     {
         public GetPlaceDto place{ get; set; }
-        public int UnixTimestamp { get; set; }
+        public long UnixTimestamp { get; set; }
 
         public long seq { get; set; }
     }
